move out of generated, code comments, other PR feedback
