+++ frontend/src/utils/validation/zodValidation.ts
@@ -13,11 +13,30 @@ type ValidationTranslator = {
   (key: string): string;
 };
 
+/**
+ * Converts a Zod issue into the validation type used by our translation keys.
+ *
+ * Zod issue codes do not map one-to-one with the validation types returned by
+ * the API. This normalizes both generated Zod validation and API validation
+ * into the same frontend vocabulary, such as:
+ *
+ * - required
+ * - not_null
+ * - invalid
+ * - min_or_max_value
+ *
+ * Custom issues generated by relational `.superRefine()` validation already
+ * contain the validation type in `issue.message`, so those pass through
+ * unchanged.
+ */
 export function getValidationTypeFromZodIssue(
   issue: z.ZodIssue,
   value: unknown,
   fieldSchema: ZodSchema,
 ): string | null {
+  // An empty string is valid input to some Zod string schemas, so use the
+  // schema's nullable/optional information to determine whether it should
+  // instead be treated as a missing required value.
   if (value === "") {
     const { nullable, optional } = unwrapSchema(fieldSchema);
 
@@ -46,6 +65,9 @@ export function getValidationTypeFromZodIssue(
       return "invalid";
 
     case z.ZodIssueCode.custom:
+      // Relational validation generated from OpenAPI uses custom issues whose
+      // message contains a specific key such as:
+      // "close_date_date_order" or "award_ceiling_numeric_order".
       return issue.message;
 
     default:
@@ -53,6 +75,19 @@ export function getValidationTypeFromZodIssue(
   }
 }
 
+/**
+ * Resolves the user-facing message for a validation failure.
+ *
+ * Translation lookup follows this order:
+ *
+ * 1. A field-specific translation, such as
+ *    `award_floor.award_ceiling_numeric_order`.
+ * 2. A generic translation for the validation type, such as `required`.
+ * 3. The original validation message as a final fallback.
+ *
+ * This allows most validation messages to be shared while still supporting
+ * field-specific wording where context matters.
+ */
 export function getTranslatedValidationMessage(
   fieldTranslations: ValidationTranslator,
   genericTranslations: ValidationTranslator,
@@ -77,6 +112,13 @@ export function getTranslatedValidationMessage(
   return fallbackMessage;
 }
 
+/**
+ * Converts Zod validation issues into the field-error structure used by forms.
+ *
+ * The complete generated schema may be validated at once so relational
+ * validation can run. The optional `field` argument allows callers to retain
+ * only one field's errors when they specifically need field-scoped results.
+ */
 export function getZodValidationErrors(
   error: z.ZodError,
   validationData: Record<string, unknown>,
@@ -88,6 +130,8 @@ export function getZodValidationErrors(
   const validationErrors: Record<string, string[]> = {};
 
   for (const issue of error.issues) {
+    // We currently support top-level form fields. Issues without a field path,
+    // or issues for fields outside this schema, are ignored here.
     const issueField = issue.path[0]?.toString();
 
     if (!issueField || !isFieldInSchema(schema, issueField)) {
@@ -118,6 +162,8 @@ export function getZodValidationErrors(
       issue.message,
     );
 
+    // A field can fail more than one rule at the same time, so preserve all
+    // translated messages rather than replacing an earlier error.
     validationErrors[issueField] = [
       ...(validationErrors[issueField] ?? []),
       message,
@@ -127,6 +173,17 @@ export function getZodValidationErrors(
   return validationErrors;
 }
 
+/**
+ * Maps API validation errors into the same field-error structure used for
+ * client-side Zod validation.
+ *
+ * Field errors use the same translation lookup as Zod issues so validation
+ * messages remain consistent whether the failure is detected before submit
+ * or returned by the API.
+ *
+ * Errors that cannot be mapped to a field in the current schema remain
+ * top-level errors rather than being attached to an unrelated form control.
+ */
 export function mapApiValidationErrors(
   response: { errors?: unknown[] | null; message?: string },
   schema: z.ZodTypeAny,
@@ -155,6 +212,8 @@ export function mapApiValidationErrors(
 
       validationErrors[field] = [...(validationErrors[field] ?? []), message];
     } else {
+      // Preserve backend messages for errors that do not belong to one of the
+      // form's fields. These are surfaced as top-level form errors.
       unmappedMessages.push(error.message ?? genericMessage);
     }
   }
